+++ R/flip_values.R
@@ -4,15 +4,15 @@
 #   Flip values                                                             ####
 
 
-#' @title Flip the values around a center value
+#' @title Flip the values around an origin
 #' @description
 #'  \Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
 #'
 #'  The values are flipped with the formula \code{x = 2 * c - x} where \code{x} is the value and \code{c} is
-#'  the center value to flip the values around.
+#'  the origin coordinate to flip the values around.
 #'
-#'  The center can be supplied as coordinates or as a function that returns coordinates. The
-#'  latter can be useful when supplying a grouped data frame and flipping around e.g. the centroid
+#'  The origin can be supplied as coordinates or as a function that returns coordinates. The
+#'  latter can be useful when supplying a grouped \code{data.frame} and flipping around e.g. the centroid
 #'  of each group.
 #'
 #'  \strong{Example}:
@@ -21,19 +21,19 @@
 #'
 #'  \code{c(5, 2, 7, 4, 3, 1)}
 #'
-#'  and the \code{center_fn = create_origin_fn(median)}
+#'  and the \code{origin_fn = create_origin_fn(median)}
 #'
 #'  Changes the values to :
 #'
 #'  \code{c(2, 5, 0, 3, 4, 6)}
 #' @author Ludvig Renbo Olsen, \email{r-pkgs@@ludvigolsen.dk}
 #' @param cols Names of columns in \code{`data`} to flip values of.
-#' @param center Coordinates of the center to flip around.
+#' @param origin Coordinates of the origin to flip around.
 #'  Must be either a single constant to use in all dimensions (columns)
 #'  or a \code{vector} with one constant per dimension.
 #'
-#'  \strong{N.B.} Ignored when \code{`center_fn`} is not \code{NULL}.
-#' @param center_fn Function for finding the center value to flip the values around.
+#'  \strong{N.B.} Ignored when \code{`origin_fn`} is not \code{NULL}.
+#' @param origin_fn Function for finding the origin value to flip the values around.
 #'  Each column will be passed as a \code{vector} in the order of \code{`cols`}.
 #'  It should return a \code{vector} with one constant per dimension.
 #'
@@ -58,7 +58,7 @@
 #'
 #'  \code{\}}
 #'
-#' @param center_col_name Name of new column with the center coordinates. If \code{NULL}, no column is added.
+#' @param origin_col_name Name of new column with the origin coordinates. If \code{NULL}, no column is added.
 #' @export
 #' @family mutate functions
 #' @inheritParams multi_mutator
@@ -85,24 +85,24 @@
 #' # Flip values of the columns
 #' flip_values(df$A)
 #' flip_values(df, cols = "A")
-#' flip_values(df, cols = "B", center = 0.3, keep_original = FALSE)
-#' flip_values(df, cols = c("A", "B"), center = c(3, 0.3),
+#' flip_values(df, cols = "B", origin = 0.3, keep_original = FALSE)
+#' flip_values(df, cols = c("A", "B"), origin = c(3, 0.3),
 #'             suffix = "",  keep_original = FALSE)
-#' flip_values(df, cols = c("A", "B"), center_fn = create_origin_fn(max))
+#' flip_values(df, cols = c("A", "B"), origin_fn = create_origin_fn(max))
 #'
 #' # Grouped by G
 #' df %>%
 #'   dplyr::group_by(G) %>%
 #'   flip_values(cols = c("A", "B"),
-#'               center_fn = create_origin_fn(median),
+#'               origin_fn = create_origin_fn(median),
 #'               keep_original = FALSE)
 #'
 #' # Plot A and flipped A
 #'
 #' # First flip A around the median and then around the value 3.
 #' df <- df %>%
-#'   flip_values(cols = "A", suffix = "_flip_median", center_col_name = NULL) %>%
-#'   flip_values(cols = "A", suffix = "_flip_3", center = 3, center_col_name = NULL)
+#'   flip_values(cols = "A", suffix = "_flip_median", origin_col_name = NULL) %>%
+#'   flip_values(cols = "A", suffix = "_flip_3", origin = 3, origin_col_name = NULL)
 #'
 #' # Plot A and A flipped around its median
 #' ggplot(df, aes(x=Index, y=A)) +
@@ -120,20 +120,20 @@
 #' }
 flip_values <- function(data,
                         cols = NULL,
-                        center = 0,
-                        center_fn = create_origin_fn(median),
+                        origin = 0,
+                        origin_fn = create_origin_fn(median),
                         suffix = "_flipped",
                         keep_original = TRUE,
-                        center_col_name = ".center") {
+                        origin_col_name = ".origin") {
 
   # Check arguments ####
   assert_collection <- checkmate::makeAssertCollection()
-  checkmate::assert_numeric(center,
+  checkmate::assert_numeric(origin,
                             min.len = 1,
                             any.missing = FALSE,
                             add = assert_collection)
-  checkmate::assert_function(center_fn, null.ok = TRUE, add = assert_collection)
-  checkmate::assert_string(center_col_name, null.ok = TRUE, add = assert_collection)
+  checkmate::assert_function(origin_fn, null.ok = TRUE, add = assert_collection)
+  checkmate::assert_string(origin_col_name, null.ok = TRUE, add = assert_collection)
   checkmate::reportAssertions(assert_collection)
   # End of argument checks ####
 
@@ -144,18 +144,18 @@ flip_values <- function(data,
     suffix = suffix,
     keep_original = keep_original,
     cols = cols,
-    center = center,
-    center_fn = center_fn,
-    center_col_name = center_col_name
+    origin = origin,
+    origin_fn = origin_fn,
+    origin_col_name = origin_col_name
   )
 }
 
 flip_mutator_method <- function(data,
                                 cols,
                                 suffix,
-                                center,
-                                center_fn,
-                                center_col_name) {
+                                origin,
+                                origin_fn,
+                                origin_col_name) {
 
   # Number of dimensions
   # Each column is a dimension
@@ -165,20 +165,20 @@ flip_mutator_method <- function(data,
   dim_vectors <- as.list(data[, cols, drop = FALSE])
 
   # Find origin if specified
-  center <- apply_coordinate_fn(
+  origin <- apply_coordinate_fn(
     dim_vectors = dim_vectors,
-    coordinates = center,
-    fn = center_fn,
+    coordinates = origin,
+    fn = origin_fn,
     num_dims = length(cols),
-    coordinate_name = "center",
-    fn_name = "center_fn",
+    coordinate_name = "origin",
+    fn_name = "origin_fn",
     dim_var_name = "cols",
     allow_len_one = TRUE
   )
 
-  # Flip around the center
+  # Flip around the origin
   dim_vectors <-
-    purrr::map2(.x = dim_vectors, .y = center, .f = ~ {
+    purrr::map2(.x = dim_vectors, .y = origin, .f = ~ {
       flip_around_(vec=.x, around = .y)
     })
 
@@ -189,8 +189,8 @@ flip_mutator_method <- function(data,
                    suffix = suffix)
 
   # Add info columns
-  if (!is.null(center_col_name)) {
-    data[[center_col_name]] <- list_coordinates(center, cols)
+  if (!is.null(origin_col_name)) {
+    data[[origin_col_name]] <- list_coordinates(origin, cols)
   }
 
   data